Moved abstract class under Php72
